+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   implemented by the driver classes) are directly dispatched by the `BackupExport::export()` and `BackupImport::import()`
   methods, and no longer by the drivers themselves;
 * added the `AbstractBackupUtility` abstract class that provides the code common to `BackupExport` and `BackupImport`;
-* `Driver::_exec()` method has been moved to the `AbstractBackupUtility` class;
+* `Driver::_exec()` method has become `AbstractBackupUtility::getProcess()`;
 * the `BackupExport::export()` and `BackupImport::import()` methods can return the filename path on success or `false`
   if the `Backup.beforeExport`/`Backup.beforeImport` events are stopped;
 * the code of `Driver::export()` and `Driver::import()` methods has been absorbed by `BackupExport::export()` and